Minor refactoring to reduce size of dxilconv.dll (#3240)
Move llvm::findScalarElement and LoopSimplifyID into separate files
Remove unused code from DxbcConverter

Reduces dxilconv.dll size from 1,225kB to 1,019kB (-17%).
Measured on x86 release build with compiler optimization for size.
